Restore sets kyma-project.io/controlled-by-provisioner to true
