+++ hack/runtime-migrator/cmd/backup-and-switch/backup.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	gardener_types "github.com/gardener/gardener/pkg/client/core/clientset/versioned/typed/core/v1beta1"
-	runtimev1 "github.com/kyma-project/infrastructure-manager/api/v1"
 	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/backup"
 	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/initialisation"
 	"github.com/kyma-project/infrastructure-manager/hack/runtime-migrator-app/internal/shoot"
@@ -13,8 +12,6 @@ import (
 	rbacv1 "k8s.io/api/rbac/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/utils/ptr"
 	"log/slog"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"slices"
@@ -120,9 +117,9 @@ func (b Backup) Do(ctx context.Context, runtimeIDs []string) error {
 		}
 
 		if b.cfg.SetControlledByKim {
-			err := setControlledByKim(ctx, b.kcpClient, runtimeID)
+			err := shoot.SetControlledByKIM(ctx, b.kcpClient, runtimeID, fieldManagerName)
 			if err != nil {
-				errMsg := fmt.Sprintf("Failed to set the rutnime to be controlled by KIM: %v", err)
+				errMsg := fmt.Sprintf("Failed to set the runtime to be controlled by KIM: %v", err)
 				b.results.ErrorOccurred(runtimeID, shootToBackup.Name, errMsg)
 				slog.Error(errMsg, "runtimeID", runtimeID)
 				continue
@@ -198,34 +195,4 @@ func labelDeprecatedCRBs(ctx context.Context, runtimeClient client.Client) ([]rb
 	}
 
 	return deprecatedCRBs, nil
-}
-
-func setControlledByKim(ctx context.Context, kcpClient client.Client, runtimeID string) error {
-	getCtx, cancelGet := context.WithTimeout(ctx, timeoutK8sOperation)
-	defer cancelGet()
-
-	key := types.NamespacedName{
-		Name:      runtimeID,
-		Namespace: "kcp-system",
-	}
-	var runtime runtimev1.Runtime
-
-	err := kcpClient.Get(getCtx, key, &runtime, &client.GetOptions{})
-	if err != nil {
-		return err
-	}
-
-	runtime.Labels["kyma-project.io/controlled-by-provisioner"] = "false"
-
-	patchCtx, cancelPatch := context.WithTimeout(ctx, timeoutK8sOperation)
-	defer cancelPatch()
-
-	runtime.Kind = "Runtime"
-	runtime.APIVersion = "infrastructuremanager.kyma-project.io/v1"
-	runtime.ManagedFields = nil
-
-	return kcpClient.Patch(patchCtx, &runtime, client.Apply, &client.PatchOptions{
-		FieldManager: fieldManagerName,
-		Force:        ptr.To(true),
-	})
 }
